moving everything related to the homepage into a module #40
