+++ src/frontend/src/app/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from './pages/login/login.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
-import { HomeComponent } from './pages/home/home.component';
+import { HomeComponent } from './home-page/home/home.component';
 import { RegisterComponent } from './pages/register/register.component';
 import { GameComponent } from './pong/game/game.component';
 import { IsInGameGuard } from './pong/game/is-in-game.guard';
