+++ src/frontend/src/app/app.module.ts
@@ -6,17 +6,17 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { LoginComponent } from './pages/login/login.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
-import { HomeComponent } from './pages/home/home.component';
+import { HomeComponent } from './home-page/home/home.component';
 import { RegisterComponent } from './pages/register/register.component';
 import { HttpClientModule } from '@angular/common/http';
-import { NavbarComponent } from './components/navbar/navbar.component';
-import { ProfilComponent } from './components/profil/profil.component';
-import { LadderComponent } from './components/ladder/ladder.component';
-import { SocialComponent } from './components/social/social.component';
-import { ChatbarComponent } from './components/chatbar/chatbar.component';
+import { NavbarComponent } from './home-page/navbar/navbar.component';
+import { ProfilComponent } from './home-page/profil/profil.component';
+import { LadderComponent } from './home-page/ladder/ladder.component';
+import { SocialComponent } from './home-page/social/social.component';
+import { ChatbarComponent } from './home-page/chatbar/chatbar.component';
 import { PongModule } from './pong/pong.module';
 import { RouterModule } from '@angular/router';
-import { PopupsComponent } from './components/popups/popups.component';
+import { PopupsComponent } from './home-page/popups/popups.component';
 
 @NgModule({
   declarations: [
