+++ src/frontend/src/app/home-page/chatbar/chatbar.component.spec.ts
