+++ src/frontend/src/app/home-page/collapse.service.spec.ts
