+++ src/frontend/src/app/home-page/home/home.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { CollapseService } from 'src/app/collapse.service';
-import { PopupsService } from 'src/app/popups.service';
+import { CollapseService } from 'src/app/home-page/collapse.service';
+import { PopupsService } from 'src/app/home-page/popups.service';
 
 @Component({
   selector: 'app-home',
