+++ src/frontend/src/app/home-page/ladder/ladder.component.spec.ts
