+++ src/frontend/src/app/home-page/ladder/ladder.component.ts
