+++ src/frontend/src/app/home-page/navbar/navbar.component.spec.ts
