+++ src/frontend/src/app/home-page/navbar/navbar.component.ts
@@ -1,6 +1,6 @@
 import { Component, Output, EventEmitter, OnInit } from '@angular/core';
-import { CollapseService } from 'src/app/collapse.service';
-import { PopupsService } from 'src/app/popups.service';
+import { CollapseService } from 'src/app/home-page/collapse.service';
+import { PopupsService } from 'src/app/home-page/popups.service';
 
 @Component({
   selector: 'app-navbar',
@@ -21,30 +21,25 @@ export class NavbarComponent implements OnInit {
   /* NAVBAR COLLAPSE */
   openSidenav(){
     this.collapseService.openNav();
-    console.log(this.collapseService.navCollapsed);
   }
 
   closeSidenav() {
     this.collapseService.closeNav();
-    console.log(this.collapseService.navCollapsed);
   }
 
   /* PROFIL POP UP */
   openProfil() {
     this.popupsService.openProfil();
-    console.log(this.popupsService.profilPopup, this.popupsService.ladderPopup, this.popupsService.socialPopup);
   }
 
   /* LADDER POP UP */
   openLadder() {
     this.popupsService.openLadder();
-    console.log(this.popupsService.profilPopup, this.popupsService.ladderPopup, this.popupsService.socialPopup);
   }
 
   /* SOCIAL POP UP */
   openSocial() {
     this.popupsService.openSocial();
-    console.log(this.popupsService.profilPopup, this.popupsService.ladderPopup, this.popupsService.socialPopup);
   }
 
 }
