+++ src/frontend/src/app/home-page/popups/popups.component.css
