+++ src/frontend/src/app/home-page/social/social.component.ts
